refactor: change printer to stringifier
