+++ src/record/export/stringifiers/stringifyAsCsv/__tests__/fixtures/header/file_schema.ts
