+++ src/record/export/stringifiers/stringifyAsCsv/__tests__/fixtures/header/schema.ts
