+++ src/record/export/stringifiers/stringifyAsCsv/__tests__/fixtures/header/subtable_schema.ts
