+++ src/record/export/stringifiers/stringifyAsCsv/__tests__/fixtures/index/common/expected.ts
