+++ src/record/export/stringifiers/stringifyAsCsv/__tests__/fixtures/index/common/input.ts
