+++ src/record/export/stringifiers/stringifyAsCsv/__tests__/fixtures/index/withEmptySubtable/index.ts
