+++ src/record/export/stringifiers/stringifyAsCsv/__tests__/fixtures/index/withEmptySubtable/input.ts
