+++ src/record/export/stringifiers/stringifyAsCsv/__tests__/fixtures/index/withFileAndAttachmentsDir/expected.ts
