+++ src/record/export/stringifiers/stringifyAsCsv/__tests__/fixtures/index/withFileAndAttachmentsDir/input.ts
