+++ src/record/export/stringifiers/stringifyAsCsv/__tests__/fixtures/index/withFileAndNoAttachmentsDir/input.ts
