+++ src/record/export/stringifiers/stringifyAsCsv/__tests__/fixtures/index/withFileAndNoAttachmentsDir/schema.ts
