+++ src/record/export/stringifiers/stringifyAsCsv/__tests__/fixtures/index/withSubtableAndFileAndAttachmentsDir/input.ts
