+++ src/record/export/stringifiers/stringifyAsCsv/__tests__/fixtures/index/withSubtableAndFileAndNoAttachmentsDir/schema.ts
