+++ src/record/export/stringifiers/stringifyAsCsv/header.ts
