+++ src/record/export/stringifiers/stringifyAsCsv/subtable.ts
